+++ src/client/tair_mc_client_api.hpp
@@ -63,7 +63,7 @@ public:
      */
     void setup_cache(int area, size_t capacity = 30);
 
-    void setup_cache(int area, size_t capacity, uint64_t expire_time);
+    void setup_cache(int area, size_t capacity, int64_t expire_time);
 
     /**
      * @brief close tairclient, release resource
@@ -89,7 +89,7 @@ public:
     int put(int area,
             const data_entry &key,
             const data_entry &data,
-            int expire,
+            int64_t expire,
             int version,
             bool fill_cache = true);
 
@@ -242,7 +242,7 @@ public:
      * @param expire&version: expire & version
      */
     int prefix_put(int area, const data_entry &pkey, const data_entry &skey,
-                   const data_entry &value, int expire, int version);
+                   const data_entry &value, int64_t expire, int version);
 
     /**
      * @brief to put multiple items with prefix key
@@ -367,18 +367,18 @@ public:
      */
     int incr(int area,
              const data_entry &key,
-             int count,
-             int *ret_count,
-             int init_value = 0,
-             int expire = 0);
+             int64_t count,
+             int64_t *ret_count,
+             int64_t init_value = 0,
+             int64_t expire = 0);
 
     // opposite to incr
     int decr(int area,
              const data_entry &key,
-             int count,
-             int *ret_count,
-             int init_value = 0,
-             int expire = 0);
+             int64_t count,
+             int64_t *ret_count,
+             int64_t init_value = 0,
+             int64_t expire = 0);
 
     /**
      *
@@ -396,9 +396,9 @@ public:
      */
     int add_count(int area,
                   const data_entry &key,
-                  int count,
-                  int *ret_count,
-                  int init_value = 0);
+                  uint64_t count,
+                  uint64_t *ret_count,
+                  uint64_t init_value = 0);
 
     /**
      * @brief set count to key's value, ignore whether this key exists or is not
@@ -412,7 +412,7 @@ public:
      *
      * @return 0 -- success, otherwise fail.
      */
-    int set_count(int area, const data_entry &key, int count, int expire = 0, int version = 0);
+    int set_count(int area, const data_entry &key, uint64_t count, int64_t expire = 0, int version = 0);
 
     /**
      * @brief Once a key is locked, it can NOT be updated(put/incr/decr), can only
@@ -449,7 +449,7 @@ public:
      */
     int expire(int area,
                const data_entry &key,
-               int expire);
+               int64_t expire);
 
     /**
      * @brief set log level
